+++ .vscode/settings.json
@@ -97,6 +97,7 @@
         "fallbacktools",
         "fetchtext",
         "ffprobe",
+        "filebytes",
         "filetree",
         "firstsecond",
         "Fmepg",
@@ -170,6 +171,7 @@
         "mcps",
         "mcpserver",
         "mdchunk",
+        "mdstringify",
         "memorystream",
         "menuitemcheckbox",
         "menuitemradio",
