+++ packages/core/src/bufferlike.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { resolveFileBytes } from "./file.js";
+import { resolveFileBytes } from "./filebytes.js";
 import { TraceOptions } from "./trace.js";
 import { fileTypeFromBuffer } from "./filetype.js";
 import { extname } from "node:path";
