+++ packages/core/src/html.ts
@@ -8,10 +8,6 @@
 import { CancellationOptions, checkCancelled } from "./cancellation.js";
 import { TraceOptions } from "./trace.js"; // Import TraceOptions for optional logging features
 import type { HTMLToMarkdownOptions, HTMLToTextOptions } from "./types.js"; // Import HTMLToTextOptions for configuring HTML to text conversion
-import { tabletojson } from "tabletojson";
-import { convert as convertToText } from "html-to-text"; // Import the convert function from html-to-text library
-import Turndown from "turndown"; // Import Turndown library for HTML to Markdown conversion
-import * as GFMPlugin from "turndown-plugin-gfm";
 
 /**
  * Converts HTML tables to JSON objects.
@@ -21,6 +17,7 @@ import * as GFMPlugin from "turndown-plugin-gfm";
  * @returns A 2D array of objects representing the table data.
  */
 export async function HTMLTablesToJSON(html: string, options?: {}): Promise<object[][]> {
+  const { tabletojson } = await import("tabletojson");
   const res = tabletojson.convert(html, options); // Convert HTML tables to JSON using tabletojson library
   return res;
 }
@@ -41,6 +38,7 @@ export async function HTMLToText(
   const { trace, cancellationToken } = options || {}; // Extract trace for logging if available
 
   try {
+    const { convert: convertToText } = await import("html-to-text"); // Import the convert function from html-to-text library
     checkCancelled(cancellationToken); // Check for cancellation token
     const text = convertToText(html, options); // Perform conversion to plain text
     return text;
@@ -65,6 +63,9 @@ export async function HTMLToMarkdown(
   const { disableGfm, trace, cancellationToken } = options || {}; // Extract trace for logging if available
 
   try {
+    const Turndown = (await import("turndown")).default; // Import Turndown library for HTML to Markdown conversion
+    const GFMPlugin = await import("turndown-plugin-gfm");
+
     checkCancelled(cancellationToken); // Check for cancellation token
     const turndown = new Turndown();
     turndown.remove("script");
