+++ packages/core/src/index.ts
@@ -65,6 +65,7 @@ export type { FetchType } from "./fetch.js";
 export * from "./fetchtext.js";
 export * from "./ffmpeg.js";
 export * from "./file.js";
+export * from "./filebytes.js";
 export * from "./filecache.js";
 export * from "./fileedits.js";
 export * from "./filetype.js";
@@ -109,6 +110,7 @@ export * from "./mcpclient.js";
 export * from "./mcpresource.js";
 export * from "./mdchunk.js";
 export * from "./mddiff.js";
+export * from "./mdstringify.js";
 export * from "./memcache.js";
 export * from "./merge.js";
 export * from "./mermaid.js";
