+++ packages/core/test/markdown.test.ts
@@ -3,7 +3,8 @@
 
 // cSpell: disable
 import { describe, test, assert } from "vitest";
-import { MarkdownStringify, splitMarkdownTextImageParts } from "../src/markdown.js";
+import { splitMarkdownTextImageParts } from "../src/markdown.js";
+import { markdownStringify } from "../src/mdstringify.js";
 import { parseTraceTree } from "../src/traceparser.js";
 
 describe("trace tree", () => {
@@ -17,15 +18,15 @@ describe("trace tree", () => {
     });
   });
   test("stringify", () => {
-    assert.strictEqual(MarkdownStringify({ a: 1 }), "\n- a: 1\n");
-    assert.strictEqual(MarkdownStringify({ a: 1, b: 2 }), "\n- a: 1\n- b: 2\n");
+    assert.strictEqual(markdownStringify({ a: 1 }), "\n- a: 1\n");
+    assert.strictEqual(markdownStringify({ a: 1, b: 2 }), "\n- a: 1\n- b: 2\n");
     assert.strictEqual(
-      MarkdownStringify({ a: "string" }, { quoteValues: true }),
+      markdownStringify({ a: "string" }, { quoteValues: true }),
       "\n- a: `string`\n",
     );
-    assert.strictEqual(MarkdownStringify([1, 2, 3]), "\n- 1\n- 2\n- 3\n");
+    assert.strictEqual(markdownStringify([1, 2, 3]), "\n- 1\n- 2\n- 3\n");
     assert.strictEqual(
-      MarkdownStringify({ a: 1 }, { headings: 0, headingLevel: 3 }),
+      markdownStringify({ a: 1 }, { headings: 0, headingLevel: 3 }),
       "\n### A\n1\n",
     );
   });
