+++ packages/vscode/build.mjs
@@ -15,7 +15,6 @@ const config = {
     "vscode",
     "pdfjs-dist",
     "@napi-rs/canvas",
-    "iconv-lite",
     "canvas",
     "sharp",
     "@xmldom/xmldom",
